tests/wpt/tests/html/editing/the-hidden-attribute/hidden-until-found-idl.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://html.spec.whatwg.org/#the-hidden-attribute">
<link rel=helph href="https://issues.chromium.org/issues/402108887">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id=target>div</div>

<script>
test(() => {
  const target = document.getElementById('target');
  target.setAttribute('hidden', 'UNTIL-FOUND');
  assert_equals(target.hidden, 'until-found');
  target.setAttribute('hidden', 'uNtIl-FoUnD');
  assert_equals(target.hidden, 'until-found');
  target.setAttribute('hidden', 'until-found');
  assert_equals(target.hidden, 'until-found');
}, 'element.hidden should return "until-found" regardless of uppercase letters.');
</script>
