LayoutTests/imported/w3c/web-platform-tests/html/editing/the-hidden-attribute/hidden-until-found-idl-expected.txt


FAIL element.hidden should return "until-found" regardless of uppercase letters. assert_equals: expected (string) "until-found" but got (boolean) true
